Split all any none tests into header and source.
